requirements.txt
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
scipy==1.17.1
matplotlib==3.10.8